+++ lib/index.d.ts
@@ -25,6 +25,8 @@ export { DEFAULT_MOVENET_LIGHTNING_URL, originFromModelUrl, resolvePoseModel, } 
 export type { PoseModelAlias, PoseModelKind, ResolvePoseModelOptions, ResolvedPoseModel, } from './models/poseModels';
 export { PoseTrackerClient } from './client';
 export type { PoseTrackerClientOptions, PreferredBackend, StartExerciseOptions } from './client';
+export type { EngineChannel } from './engineChannel';
+export { V4_ONLY_EXERCISE_IDS, normalizeEngineChannel, requiresEngineV4 } from './engineChannel';
 export type { ColdStartMode, PreloadOptions } from './types/preload';
 export type { SkeletonAnglesStyle, SkeletonCirclesStyle, SkeletonDefinition, SkeletonLinesStyle, } from './types/skeleton';
 export { DEFAULT_SKELETON_DEFINITION } from './types/skeleton';